requirements.txt
flask-login
psycopg2-binary
sqlalchemy
flask-session
gunicorn
